fix: snapshot file blobs on each edit so line-level blame works for agent commits (#18)
* docs: update README and CONTRIBUTING to match current state

Closes #13

- Update tool count from 10 to 15 across README tagline and features list
- Add all 15 supported tools to the features bullet (was missing Amp,
  Continue, Factory Droid, Junie, Kiro)
- Mark Kiro, Continue, and Factory Droid with agent hooks support in the
  supported tools table (hooks added in recent releases)
- Update agent lifecycle hooks section to list all 7 hooked tools
- Add 5 missing config keys (kiro, continue, droid, junie, amp) to the
  configuration reference
- Update CONTRIBUTING command count from 17 to 23
- Add 5 missing tool module entries to CONTRIBUTING project overview table
- Update CI description: add Windows testing, security audit, clippy -D
  warnings, clarify Docker containers

Made-with: Cursor

* fix: snapshot file blobs on each edit so line-level blame works for agent commits

Previously, file_snapshots (git blob hashes) were only written by the
`stop` hook at end-of-turn. When an agent committed mid-session, the
anchor was built before `stop` ran, so line_attributions was always
empty — producing "file-level attribution only" in `oobo blame`.

Now `after-tool-use` / `after-file-edit` snapshots the blob immediately
after recording the edited file path. This ensures `enrich_commit` has
blob data available whenever a commit happens during the session.

The `stop` handler still re-snapshots all files as a final pass, which
is safe (later writes overwrite earlier ones for the same path).

Made-with: Cursor

* fix: use committed blob for agent commits to avoid stale snapshot misattribution

When an agent edits a file multiple times in one session and reverts
some changes, the after-tool-use snapshot captures an intermediate
state. The 3-way diff then sees the difference between the stale
snapshot and the committed blob as human edits even though no human
was involved.

For agent-authored commits (non-interactive), the committed blob is
by definition the agents final output. This patch replaces stale
snapshot blobs with the committed blob hash before building the
anchor, so compute_precise_attribution correctly attributes all
lines as AI.

Assisted (human + agent) commits still use the original snapshot
for the 3-way diff, preserving the ability to separate human edits
from agent edits.

Made-with: Cursor

* fix: check tool_input.path as fallback for file_path in after-tool-use

Cursor sends the file path as tool_input.path for Write events, not
tool_input.file_path. The hook handler only checked file_path, so
edited files were tracked (via record_edited_file which uses a
separate path) but never snapshotted during the session — causing
empty line_attributions in the anchor.

Made-with: Cursor

* fix: resolve clippy warnings and update vulnerable dependencies

- Replace sort_by with sort_by_key where applicable (clippy::unnecessary_sort_by)
- Remove unnecessary `as i64` cast in aider.rs (clippy::unnecessary_cast)
- Collapse if-blocks inside match arms into match guards (clippy::collapsible_if)
- Update rustls-webpki 0.103.10 -> 0.103.12 (RUSTSEC-2026-0098, RUSTSEC-2026-0099)
- Update rand 0.8.5 -> 0.8.6, 0.9.2 -> 0.9.4 (RUSTSEC-2026-0097)

Made-with: Cursor

* fix(ci): use workspace-relative CARGO_HOME for Docker container jobs

The alpine container job fails because $HOME differs between the
"Install Rust" and "Build and test" steps, so `. "$HOME/.cargo/env"`
cannot find the file. Pin CARGO_HOME to the workspace directory so
the path is stable across steps.

Made-with: Cursor

* fix(ci): drop .cargo/env sourcing in Docker jobs, rely on GITHUB_PATH

The install step already adds cargo bin to GITHUB_PATH (expanded at
runtime with the correct HOME), so cargo is on PATH in subsequent
steps without needing to source the env file.

Made-with: Cursor

+++ src/tools/aider.rs
@@ -163,7 +163,7 @@ fn parse_aider_timestamp(s: &str) -> Option<i64> {
     let s: u32 = time_parts[2].parse().ok()?;
 
     let days = days_from_epoch(y, mo, d)?;
-    let secs = days as i64 * 86400 + h as i64 * 3600 + mi as i64 * 60 + s as i64;
+    let secs = days * 86400 + i64::from(h) * 3600 + i64::from(mi) * 60 + i64::from(s);
     Some(secs * 1000)
 }
